re-organize dUPE scripts and simplify the path to them
